backend/requirements.txt
flask>=1.0.2
pyjwt>=1.4.2